requirements.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.1.3
djangorestframework==3.14.0
filelock==3.8.0
platformdirs==2.5.3
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
virtualenv==20.16.6
